@@ -11,10 +11,10 @@ description: Change the world.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.3.0-alpha.1+17
+version: 1.3.0-beta.1+19
 
 environment:
-  sdk: ">=2.13.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -33,24 +33,32 @@ dependencies:
   file_picker: ^4.5.1
   phosphor_flutter: ^1.4.0
   replay_bloc: ^0.2.1
-  share_plus: ^3.1.0
-  package_info_plus: ^1.4.0
+  share_plus: ^4.0.4
+  package_info_plus: ^1.4.2
   idb_shim: ^2.0.1
   flutter_launcher_icons: ^0.9.2
-  go_router: ^3.0.4
+  go_router: ^3.0.7
   animations: ^2.0.2
-  collection: ^1.15.0
+  xml: ^5.4.0
+  collection: ^1.16.0
   bloc_concurrency: ^0.2.0
-  camera: ^0.9.4+16
+  camera: ^0.9.4+19
   camera_windows: ^0.1.0
-  window_manager: ^0.2.0
+  window_manager: ^0.2.1
   flex_color_scheme: ^4.2.0
+  freezed_annotation: ^2.0.1
+  json_annotation: ^4.4.0
+  path: ^1.8.1
 
 dev_dependencies:
-  flutter_native_splash: ^2.1.0
-  flutter_lints: ^1.0.4
+  flutter_native_splash: ^2.1.6
+  flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
+  analyzer: ^4.0.0
+  freezed: ^2.0.2
+  build_runner: ^2.1.10
+  json_serializable: ^6.1.6
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
