@@ -1,23 +1,24 @@
 import 'package:butterfly/api/open_image.dart';
 import 'package:butterfly/bloc/document_bloc.dart';
-import 'package:butterfly/cubits/selection.dart';
 import 'package:butterfly/dialogs/elements/general.dart';
-import 'package:butterfly/models/elements/image.dart';
+import 'package:butterfly/widgets/context_menu.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../widgets/exact_slider.dart';
+import '../../models/element.dart';
+import '../../renderers/renderer.dart';
+import '../constraints.dart';
 
 class ImageElementDialog extends StatefulWidget {
-  final int index;
+  final ImageRenderer renderer;
   final VoidCallback close;
   final Offset position;
 
   const ImageElementDialog(
       {Key? key,
-      required this.index,
+      required this.renderer,
       required this.close,
       required this.position})
       : super(key: key);
@@ -27,44 +28,65 @@ class ImageElementDialog extends StatefulWidget {
 }
 
 class _ImageElementDialogState extends State<ImageElementDialog> {
-  late ImageElement element;
+  late ImageElement element, newElement;
+  late DocumentBloc bloc;
 
   @override
   void initState() {
     super.initState();
-    var bloc = context.read<DocumentBloc>();
-    element = (bloc.state as DocumentLoadSuccess).document.content[widget.index]
-        as ImageElement;
+    element = widget.renderer.element;
+    newElement = element;
+    bloc = context.read<DocumentBloc>();
   }
 
   void _changeElement() {
-    context.read<DocumentBloc>().add(ElementChanged(widget.index, element));
-    context.read<SelectionCubit>().change(element);
+    bloc.add(ElementChanged(element, newElement));
+    element = newElement;
   }
 
   @override
   Widget build(BuildContext context) {
+    String constraints;
+    if (element.constraints is ScaledElementConstraints) {
+      constraints = AppLocalizations.of(context)!.scaledConstraints;
+    } else if (element.constraints is FixedElementConstraints) {
+      constraints = AppLocalizations.of(context)!.fixedConstraints;
+    } else if (element.constraints is DynamicElementConstraints) {
+      constraints = AppLocalizations.of(context)!.dynamicConstraints;
+    } else {
+      constraints = AppLocalizations.of(context)!.none;
+    }
     return GeneralElementDialog(
       close: widget.close,
-      index: widget.index,
+      renderer: widget.renderer,
       position: widget.position,
       children: [
-        ExactSlider(
-          header: Text(AppLocalizations.of(context)!.scale),
-          value: element.scale,
-          min: 0.1,
-          max: 5,
-          defaultValue: 1,
-          onChanged: (value) {
-            setState(() => element = element.copyWith(scale: value));
+        ListTile(
+          title: Text(AppLocalizations.of(context)!.constraints),
+          leading: const Icon(PhosphorIcons.selectionLight),
+          subtitle: Text(constraints),
+          onTap: () {
+            widget.close();
+            showContextMenu(
+                context: context,
+                position: widget.position,
+                builder: (context, close) => ConstraintsContextMenu(
+                    position: widget.position,
+                    enableScaled: true,
+                    initialConstraints: element.constraints,
+                    close: close,
+                    onChanged: (constraints) {
+                      close();
+                      newElement = element.copyWith(constraints: constraints);
+                      _changeElement();
+                    }));
           },
-          onChangeEnd: (value) => _changeElement(),
         ),
         ListTile(
           title: Text(AppLocalizations.of(context)!.export),
           leading: const Icon(PhosphorIcons.exportLight),
           onTap: () {
-            openImage(element.pixels);
+            openImage(newElement.pixels);
           },
         ),
       ],
