@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:flutter/services.dart';
 import 'package:window_manager/window_manager.dart';
 
-bool isFullscreen = false;
+late bool isFullscreen;
 Future<bool> isFullScreen() async {
   if (Platform.isAndroid || Platform.isIOS) {
     return isFullscreen;
@@ -13,20 +13,17 @@ Future<bool> isFullScreen() async {
 }
 
 void setup() {
-  if (Platform.isAndroid || Platform.isIOS) {
-    SystemChrome.setSystemUIChangeCallback((systemOverlaysAreVisible) async =>
-        isFullscreen = systemOverlaysAreVisible);
-  }
+  isFullscreen = false;
 }
 
 Future<void> setFullScreen(bool fullScreen) async {
   if (Platform.isAndroid || Platform.isIOS) {
     if (fullScreen) {
-      await SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual,
-          overlays: []);
+      await SystemChrome.setEnabledSystemUIMode(SystemUiMode.immersiveSticky);
       isFullscreen = true;
     } else {
-      await SystemChrome.restoreSystemUIOverlays();
+      await SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual,
+          overlays: SystemUiOverlay.values);
       isFullscreen = false;
     }
   } else {
