@@ -3,77 +3,118 @@ import 'dart:ui' as ui;
 import 'package:bloc_concurrency/bloc_concurrency.dart';
 import 'package:butterfly/api/file_system.dart';
 import 'package:butterfly/cubits/transform.dart';
-import 'package:butterfly/models/baked_viewport.dart';
+import 'package:butterfly/models/background.dart';
+import 'package:butterfly/models/current_index.dart';
 import 'package:butterfly/models/document.dart';
-import 'package:butterfly/models/elements/element.dart';
-import 'package:butterfly/models/painters/painter.dart';
+import 'package:butterfly/models/painter.dart';
 import 'package:butterfly/models/palette.dart';
-import 'package:butterfly/models/properties/hand.dart';
+import 'package:butterfly/models/viewport.dart';
 import 'package:butterfly/models/waypoint.dart';
 import 'package:butterfly/view_painter.dart';
 import 'package:collection/collection.dart';
 import 'package:equatable/equatable.dart';
 import 'package:replay_bloc/replay_bloc.dart';
 
+import '../handlers/handler.dart';
 import '../models/area.dart';
-import '../models/elements/image.dart';
+import '../models/element.dart';
+import '../models/property.dart';
+import '../renderers/renderer.dart';
 
 part 'document_event.dart';
 part 'document_state.dart';
 
 class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
-  DocumentBloc(AppDocument initial, String? path)
-      : super(DocumentLoadSuccess(initial, path: path)) {
+  DocumentBloc(AppDocument initial, String? path,
+      Renderer<Background> background, List<Renderer<PadElement>> renderer)
+      : super(DocumentLoadSuccess(initial,
+            path: path,
+            cameraViewport: CameraViewport.unbaked(background, renderer))) {
     _init();
   }
 
   void _init() {
     on<ElementsCreated>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
+        final renderers =
+            event.elements.map((e) => Renderer.fromInstance(e)).toList();
+        await Future.wait(
+            renderers.map((e) async => await e.setup(current.document)));
         return _saveDocument(
             current.copyWith(
                 document: current.document.copyWith(
                     content: (List.from(current.document.content)
                       ..addAll(event.elements)))),
-            List<PadElement>.from(current.elements)..addAll(event.elements));
+            renderers);
       }
     });
     on<ElementsReplaced>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
-        var elements = List<PadElement>.from(current.elements);
+        final current = state as DocumentLoadSuccess;
+        var renderers = List<Renderer<PadElement>>.from(current.renderers);
         event.replacedElements.forEach((index, element) {
+          final current = element.map((e) => Renderer.fromInstance(e));
           if (index == null) {
-            elements.addAll(element);
+            renderers.addAll(current);
           } else {
-            elements.removeAt(index);
-            elements.insertAll(index, element);
+            renderers.removeAt(index);
+            renderers.insertAll(index, current);
           }
         });
         return _saveDocument(
             current.copyWith(
-                document: current.document.copyWith(content: elements)),
+              document: current.document
+                  .copyWith(content: renderers.map((e) => e.element).toList()),
+            ),
             null);
       }
     });
     on<ElementChanged>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
+        final renderers = <Renderer<PadElement>>[];
+        Renderer<PadElement>? oldRenderer, newRenderer;
+        for (var renderer in current.renderers) {
+          if (renderer.element == event.old) {
+            newRenderer = Renderer.fromInstance(event.updated);
+            await newRenderer.setup(current.document);
+            oldRenderer = renderer;
+            renderers.add(newRenderer);
+          } else {
+            renderers.add(renderer);
+          }
+        }
+        final index = current.document.content.indexOf(event.old);
+        if (index < 0) return;
         return _saveDocument(
-            current.copyWith(
-              document: current.document.copyWith(
-                  content: (List.from(current.document.content)
-                    ..[event.index] = event.element)),
-            ),
-            null);
+                current.copyWith(
+                    document: current.document.copyWith(
+                        content: (current.document.content)
+                          ..[index] = event.updated),
+                    cameraViewport: current.cameraViewport.unbake(renderers)),
+                null)
+            .then((value) async {
+          if (oldRenderer == null || newRenderer == null) return;
+          if (await current.currentIndex.handler
+              .onRendererUpdated(current.document, oldRenderer, newRenderer)) {
+            add(const IndexRefreshed());
+          }
+        });
       }
     });
     on<ElementsRemoved>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(
             current.copyWith(
+                cameraViewport: current.cameraViewport.withUnbaked(
+                  current.renderers
+                      .where((element) => !event.elements.contains(
+                            element.element,
+                          ))
+                      .toList(),
+                ),
                 document: current.document.copyWith(
                     content: List.from(current.document.content)
                       ..removeWhere(
@@ -83,30 +124,46 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
     });
     on<DocumentDescriptorChanged>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(current.copyWith(
-            document: current.document
-                .copyWith(name: event.name, description: event.description)));
+            document: current.document.copyWith(
+                name: event.name ?? current.document.name,
+                description:
+                    event.description ?? current.document.description)));
       }
     });
 
     on<DocumentPaletteChanged>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(current.copyWith(
             document: current.document.copyWith(palettes: event.palette)));
       }
     });
     on<CurrentPainterChanged>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        return _saveDocument((state as DocumentLoadSuccess).copyWith(
-            currentPainterIndex: event.painter,
-            removeCurrentPainterIndex: event.painter == null));
+        final handler = Handler.fromBloc(this, event.painter);
+        emit((state as DocumentLoadSuccess).copyWith(
+            currentIndex: CurrentIndex(event.painter ?? -1, handler),
+            removeCurrentIndex: event.painter == null));
+      }
+    });
+    on<IndexRefreshed>((event, emit) async {
+      if (state is DocumentLoadSuccess) {
+        final current = state as DocumentLoadSuccess;
+        final index = current.currentIndex;
+        emit(current.copyWith(
+          currentIndex: index.copyWith(
+              foregrounds: index.handler
+                  .createForegrounds(current.document, current.currentArea),
+              selections: index.handler
+                  .createSelections(current.document, current.currentArea)),
+        ));
       }
     });
     on<PainterCreated>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(current.copyWith(
             document: current.document.copyWith(
                 painters: List.from(current.document.painters)
@@ -115,7 +172,7 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
     });
     on<PainterChanged>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(current.copyWith(
             document: current.document.copyWith(
                 painters: List.from(current.document.painters)
@@ -124,7 +181,7 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
     });
     on<PainterRemoved>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(current.copyWith(
             document: current.document.copyWith(
                 painters: List.from(current.document.painters)
@@ -133,7 +190,7 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
     });
     on<PainterReordered>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         var painters = List<Painter>.from(current.document.painters);
         var oldIndex = event.oldIndex;
         var newIndex = event.newIndex;
@@ -142,27 +199,34 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
         }
         final item = painters.removeAt(oldIndex);
         painters.insert(newIndex, item);
+        final index = CurrentIndex(
+            oldIndex == current.currentIndex.index
+                ? newIndex
+                : newIndex == current.currentIndex.index
+                    ? oldIndex
+                    : current.currentIndex.index,
+            Handler.fromBloc(this));
         return _saveDocument(current.copyWith(
             document: current.document.copyWith(painters: painters),
-            currentPainterIndex: oldIndex == current.currentPainterIndex
-                ? newIndex
-                : newIndex == current.currentPainterIndex
-                    ? oldIndex
-                    : current.currentPainterIndex));
+            currentIndex: index));
       }
     });
     on<DocumentBackgroundChanged>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
+        final Renderer<Background> background =
+            Renderer.fromInstance(event.background);
+        await background.setup(current.document);
         return _saveDocument(current.copyWith(
+            cameraViewport: current.cameraViewport.withBackground(background),
             document: current.document.copyWith(
-                background: event.background,
-                removeBackground: event.background == null)));
+              background: event.background,
+            )));
       }
     });
     on<WaypointCreated>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(current.copyWith(
             document: current.document.copyWith(
                 waypoints: List<Waypoint>.from(current.document.waypoints)
@@ -171,7 +235,7 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
     });
     on<WaypointRemoved>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(current.copyWith(
             document: current.document.copyWith(
                 waypoints: List<Waypoint>.from(current.document.waypoints)
@@ -180,7 +244,7 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
     });
     on<HandPropertyChanged>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(current.copyWith(
             document: current.document.copyWith(handProperty: event.property)));
       }
@@ -188,7 +252,7 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
 
     on<LayerRenamed>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(current.copyWith(
             document: current.document.copyWith(
                 content: List<PadElement>.from(current.document.content)
@@ -201,7 +265,7 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
 
     on<LayerRemoved>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(
             current.copyWith(
                 document: current.document.copyWith(
@@ -215,7 +279,7 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
 
     on<LayerElementsDeleted>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(
             current.copyWith(
                 document: current.document.copyWith(
@@ -228,7 +292,7 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
 
     on<LayerVisibilityChanged>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         var invisibleLayers = List<String>.from(current.invisibleLayers);
         var isVisible = current.isLayerVisible(event.name);
         if (isVisible) {
@@ -238,17 +302,22 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
         }
         return _saveDocument(
             current.copyWith(
-                removeBakedViewport: isVisible,
+                cameraViewport: isVisible
+                    ? current.cameraViewport.unbake(current.renderers)
+                    : null,
                 invisibleLayers: invisibleLayers),
-            isVisible ? null : List<PadElement>.from(current.elements)
+            isVisible
+                ? null
+                : List<Renderer<PadElement>>.from(current.renderers)
               ?..addAll(current.document.content
-                  .where((e) => e.layer == event.name)));
+                  .where((e) => e.layer == event.name)
+                  .map((e) => Renderer.fromInstance(e))));
       }
     });
 
     on<CurrentLayerChanged>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         return _saveDocument(current.copyWith(
           currentLayer: event.name,
         ));
@@ -257,7 +326,7 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
 
     on<ElementsLayerChanged>((event, emit) async {
       if (state is DocumentLoadSuccess) {
-        var current = state as DocumentLoadSuccess;
+        final current = state as DocumentLoadSuccess;
         var content = List<PadElement>.from(current.document.content);
         for (var element in event.elements) {
           var i = current.document.content.indexOf(element);
@@ -279,73 +348,65 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
       }
       final eq = const ListEquality().equals;
 
-      var elements = current.elements;
+      var renderers = current.renderers;
       var recorder = ui.PictureRecorder();
       var size = event.viewportSize;
       var canvas = ui.Canvas(recorder);
-      var last = current.bakedViewport;
+      var last = current.cameraViewport;
       var invisibleLayers = current.invisibleLayers;
-      var reset = last == null ||
-          last.width != size.width ||
-          last.height != size.height ||
+      var reset = last.width != size.width.round() ||
+          last.height != size.height.round() ||
           last.x != event.cameraTransform.position.dx ||
           last.y != event.cameraTransform.position.dy ||
           last.scale != event.cameraTransform.size;
-      if (elements.isEmpty && !reset) return;
+      if (renderers.isEmpty && !reset) return;
       if (reset) {
-        elements = current.document.content
-            .where((element) => !invisibleLayers.contains(element.layer))
+        renderers = renderers
+            .where(
+                (element) => !invisibleLayers.contains(element.element.layer))
             .toList();
       }
-
-      var images = Map<ImageElement, ui.Image>.from(last?.images ?? {});
-      // Bake images if not baked yet
-      for (var element in elements) {
-        if (element is ImageElement && !images.containsKey(element)) {
-          images[element] = await loadImage(element);
-        }
-      }
-
       ViewPainter(
         current.document,
-        elements: elements,
         transform: event.cameraTransform,
-        bakedViewport: reset ? null : last,
+        cameraViewport: reset ? current.cameraViewport.unbake(renderers) : last,
         renderBackground: false,
-        images: images,
       ).paint(canvas, event.viewportSize);
 
       var picture = recorder.endRecording();
       var newImage =
           await picture.toImage((size.width).ceil(), (size.height).ceil());
       current = state as DocumentLoadSuccess;
-      var currentElements = current.elements;
+      var currentElements = current.renderers;
       if (reset) {
-        currentElements = current.document.content
-            .where((element) => !invisibleLayers.contains(element.layer))
+        currentElements = renderers
+            .where(
+                (element) => !invisibleLayers.contains(element.element.layer))
             .toList();
       }
-      if (!eq(elements, currentElements)) return;
-      if (last != current.bakedViewport) return;
+      if (!eq(renderers, currentElements)) return;
+      if (last != current.cameraViewport) return;
 
       emit(current.copyWith(
-          bakedViewport: BakedViewport(
+          cameraViewport: current.cameraViewport.bake(
               height: size.height.round(),
               width: size.width.round(),
               scale: event.cameraTransform.size,
               x: event.cameraTransform.position.dx,
               y: event.cameraTransform.position.dy,
               image: newImage,
+              bakedElements: renderers,
               unbakedElements: [])));
     }, transformer: restartable());
     on<ImageUnbaked>((event, emit) {
-      var current = state;
+      final current = state;
       if (current is DocumentLoadSuccess) {
-        emit(current.copyWith(removeBakedViewport: true));
+        emit(current.copyWith(
+            cameraViewport: current.cameraViewport.unbake(current.renderers)));
       }
     });
     on<TemplateCreated>((event, emit) {
-      var current = state;
+      final current = state;
       if (current is! DocumentLoadSuccess) return;
       TemplateFileSystem.fromPlatform().createTemplate(current.document);
 
@@ -354,77 +415,86 @@ class DocumentBloc extends ReplayBloc<DocumentEvent, DocumentState> {
       }
     });
     on<DocumentPathChanged>((event, emit) {
-      var current = state;
+      final current = state;
       if (current is! DocumentLoadSuccess) return;
       emit(current.copyWith(path: event.path));
     });
-    on<AreaCreated>((event, emit) {
-      var current = state;
+    on<AreaCreated>((event, emit) async {
+      final current = state;
       if (current is! DocumentLoadSuccess) return;
-      emit(current.copyWith(
-          document: current.document.copyWith(
-              areas: List<Area>.from(current.document.areas)
-                ..add(event.area))));
+      final currentDocument = current.document.copyWith(
+          areas: List<Area>.from(current.document.areas)..add(event.area));
+      emit(current.copyWith(document: currentDocument));
+      for (var element in current.renderers) {
+        if (await element.onAreaUpdate(currentDocument, event.area)) {
+          _repaint();
+        }
+      }
     });
-    on<AreaRemoved>((event, emit) {
-      var current = state;
+    on<AreaRemoved>((event, emit) async {
+      final current = state;
       if (current is! DocumentLoadSuccess) return;
-      emit(current.copyWith(
-          document: current.document.copyWith(
-              areas: List<Area>.from(current.document.areas)
-                ..removeAt(event.index))));
+      final areas = List<Area>.from(current.document.areas);
+      final area = areas.removeAt(event.index);
+      final currentDocument = current.document.copyWith(areas: areas);
+      emit(current.copyWith(document: currentDocument));
+      for (var element in current.renderers) {
+        if (element.area == area &&
+            await element.onAreaUpdate(currentDocument, null)) {
+          _repaint();
+        }
+      }
     });
-    on<AreaChanged>((event, emit) {
-      var current = state;
+    on<AreaChanged>((event, emit) async {
+      final current = state;
       if (current is! DocumentLoadSuccess) return;
-      emit(current.copyWith(
-          document: current.document.copyWith(
-              areas: List<Area>.from(current.document.areas)
-                ..[event.index] = event.area)));
+      final areas = List<Area>.from(current.document.areas);
+      final area = areas[event.index];
+      final currentDocument =
+          current.document.copyWith(areas: areas..[event.index] = event.area);
+      emit(current.copyWith(document: currentDocument));
+      for (var element in current.renderers) {
+        if (element.area == area) {
+          if (await element.onAreaUpdate(currentDocument, event.area)) {
+            _repaint();
+          }
+        }
+      }
     });
     on<CurrentAreaChanged>((event, emit) {
-      var current = state;
+      final current = state;
       if (current is! DocumentLoadSuccess) return;
       emit(current.copyWith(currentAreaIndex: event.area));
     });
   }
 
-  @override
-  void onChange(Change<DocumentState> change) {
-    // Always call super.onChange with the current change
-    super.onChange(change);
-
-    // Custom onChange logic goes here
-
-    var last = change.currentState;
-    var current = change.nextState;
-    if (last is! DocumentLoadSuccess || current is! DocumentLoadSuccess) return;
-    if (last.path != null &&
-        last.bakedViewport?.image != current.bakedViewport?.image) {
-      last.bakedViewport?.dispose();
+  Future<void> _saveDocument(DocumentLoadSuccess current,
+      [List<Renderer<PadElement>>? unbakedElements = const []]) async {
+    var elements = current.cameraViewport.unbakedElements;
+    if (unbakedElements != null) {
+      for (var renderer in unbakedElements) {
+        await renderer.setup(current.document);
+      }
+      elements = List<Renderer<PadElement>>.from(elements)
+        ..addAll(unbakedElements);
+    }
+    var nextState = current.copyWith(
+        document: current.document.copyWith(updatedAt: DateTime.now()),
+        cameraViewport: unbakedElements == null
+            ? current.cameraViewport.unbake(elements)
+            : current.cameraViewport.withUnbaked(elements));
+    if (nextState.path != null) {
+      emit(nextState);
+    }
+    var path = await nextState.save();
+    if (nextState.path == null) {
+      emit(nextState.copyWith(path: path));
     }
   }
 
-  Future<void> _saveDocument(DocumentLoadSuccess current,
-      [List<PadElement>? unbakedElements = const []]) async {
-    var elements = current.elements;
-    if (current.bakedViewport != null &&
-        unbakedElements != null &&
-        current.bakedViewport != null) {
-      elements = List<PadElement>.from(elements)..addAll(unbakedElements);
-    }
-    current = current.copyWith(
-        document: current.document.copyWith(updatedAt: DateTime.now()),
-        removeBakedViewport: unbakedElements == null,
-        bakedViewport: unbakedElements == null
-            ? current.bakedViewport
-            : current.bakedViewport?.withUnbaked(elements));
-    if (current.path != null) {
-      emit(current);
-    }
-    var path = await current.save();
-    if (current.path == null) {
-      emit(current.copyWith(path: path));
-    }
+  void _repaint() {
+    final current = state;
+    if (current is! DocumentLoadSuccess) return;
+    emit(current.copyWith(cameraViewport: current.cameraViewport.unbake()));
   }
 }
