@@ -49,6 +49,7 @@ class AreaContextMenu extends StatelessWidget {
                 } else {
                   bloc.add(CurrentAreaChanged(index));
                 }
+                bloc.add(const IndexRefreshed());
               },
             ),
             ListTile(
@@ -56,8 +57,8 @@ class AreaContextMenu extends StatelessWidget {
               title: Text(AppLocalizations.of(context)!.name),
               subtitle: Text(area.name),
               onTap: () {
-                var _nameController = TextEditingController(text: area.name);
-                var formKey = GlobalKey<FormState>();
+                final nameController = TextEditingController(text: area.name);
+                final formKey = GlobalKey<FormState>();
                 close();
                 showDialog(
                   context: context,
@@ -76,7 +77,7 @@ class AreaContextMenu extends StatelessWidget {
                           return null;
                         },
                         decoration: const InputDecoration(filled: true),
-                        controller: _nameController,
+                        controller: nameController,
                         autofocus: true,
                       ),
                       actions: [
@@ -91,7 +92,7 @@ class AreaContextMenu extends StatelessWidget {
                             bloc.add(
                               AreaChanged(
                                 index,
-                                area.copyWith(name: _nameController.text),
+                                area.copyWith(name: nameController.text),
                               ),
                             );
                           },
@@ -109,13 +110,15 @@ class AreaContextMenu extends StatelessWidget {
                 close();
                 var bloc = context.read<DocumentBloc>();
                 showDialog(
-                    builder: (context) => ImageExportDialog(
-                          bloc: bloc,
-                          width: area.width.round(),
-                          height: area.height.round(),
-                          x: area.position.dx,
-                          y: area.position.dy,
-                          scale: 1,
+                    builder: (context) => BlocProvider.value(
+                          value: bloc,
+                          child: ImageExportDialog(
+                            width: area.width.round(),
+                            height: area.height.round(),
+                            x: area.position.dx,
+                            y: area.position.dy,
+                            scale: 1,
+                          ),
                         ),
                     context: context);
               },
