@@ -1,18 +1,19 @@
 import 'package:butterfly/bloc/document_bloc.dart';
-import 'package:butterfly/cubits/editing.dart';
+import 'package:butterfly/dialogs/area/context.dart';
 import 'package:butterfly/dialogs/layer.dart';
-import 'package:butterfly/models/elements/element.dart';
+import 'package:butterfly/models/element.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 
 import '../../cubits/transform.dart';
+import '../../renderers/renderer.dart';
 import '../../widgets/context_menu.dart';
 import '../background/context.dart';
 
 class GeneralElementDialog extends StatelessWidget {
-  final int index;
+  final Renderer<PadElement> renderer;
   final VoidCallback close;
   final Offset position;
   final List<Widget> children;
@@ -21,23 +22,18 @@ class GeneralElementDialog extends StatelessWidget {
       {Key? key,
       this.children = const [],
       required this.position,
-      required this.index,
+      required this.renderer,
       required this.close})
       : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return BlocBuilder<DocumentBloc, DocumentState>(builder: (context, state) {
+      if (state is! DocumentLoadSuccess) return Container();
       var bloc = context.read<DocumentBloc>();
-      var editingCubit = context.read<EditingCubit>();
-      if (state is! DocumentLoadSuccess ||
-          index < 0 ||
-          index >= state.document.content.length) {
-        return Container();
-      }
-      var element = state.document.content[index];
       return Column(mainAxisSize: MainAxisSize.min, children: [
-        generateHeader(element),
+        _GeneralElementDialogHeader(
+            close: close, renderer: renderer, position: position),
         const Divider(),
         Flexible(
             child: ListView(
@@ -45,15 +41,15 @@ class GeneralElementDialog extends StatelessWidget {
           children: [
             ...children,
             if (children.isNotEmpty) const Divider(),
-            element.layer.isEmpty
+            renderer.element.layer.isEmpty
                 ? ListTile(
                     title: Text(AppLocalizations.of(context)!.layer),
                     leading: const Icon(PhosphorIcons.squaresFourLight),
                     subtitle: Text(AppLocalizations.of(context)!.notSet),
                     onTap: () {
                       close();
-                      var _nameController =
-                          TextEditingController(text: element.layer);
+                      var nameController =
+                          TextEditingController(text: renderer.element.layer);
                       showDialog(
                           context: context,
                           useRootNavigator: true,
@@ -61,11 +57,13 @@ class GeneralElementDialog extends StatelessWidget {
                                 title: Text(
                                     AppLocalizations.of(context)!.enterLayer),
                                 content: TextField(
-                                    controller: _nameController,
-                                    autofocus: true,
-                                    decoration: const InputDecoration(
+                                  controller: nameController,
+                                  autofocus: true,
+                                  decoration: InputDecoration(
                                       filled: true,
-                                    )),
+                                      hintText:
+                                          AppLocalizations.of(context)!.name),
+                                ),
                                 actions: [
                                   TextButton(
                                     child: Text(
@@ -78,9 +76,9 @@ class GeneralElementDialog extends StatelessWidget {
                                         Text(AppLocalizations.of(context)!.ok),
                                     onPressed: () {
                                       bloc.add(ElementChanged(
-                                          index,
-                                          element.copyWith(
-                                              layer: _nameController.text)));
+                                          renderer.element,
+                                          renderer.element.copyWith(
+                                              layer: nameController.text)));
                                       Navigator.of(context).pop();
                                     },
                                   ),
@@ -89,11 +87,11 @@ class GeneralElementDialog extends StatelessWidget {
                     })
                 : ListTile(
                     leading: const Icon(PhosphorIcons.squaresFourLight),
-                    title: Text(element.layer),
+                    title: Text(renderer.element.layer),
                     trailing: IconButton(
                       icon: const Icon(PhosphorIcons.trashLight),
-                      onPressed: () => bloc.add(
-                          ElementChanged(index, element.copyWith(layer: ''))),
+                      onPressed: () => bloc.add(ElementChanged(renderer.element,
+                          renderer.element.copyWith(layer: ''))),
                     ),
                     onTap: () {
                       close();
@@ -101,15 +99,16 @@ class GeneralElementDialog extends StatelessWidget {
                           context: context,
                           builder: (context) => BlocProvider.value(
                               value: bloc,
-                              child: LayerDialog(layer: element.layer)));
+                              child:
+                                  LayerDialog(layer: renderer.element.layer)));
                     }),
             ListTile(
               title: Text(AppLocalizations.of(context)!.move),
               leading: const Icon(PhosphorIcons.arrowsOutCardinalLight),
               onTap: () {
                 // Remove the element from the document
-                bloc.add(ElementsRemoved([element]));
-                editingCubit.move(element);
+                bloc.add(ElementsRemoved([renderer.element]));
+                state.fetchHand()?.move(context, renderer);
                 close();
               },
             ),
@@ -117,7 +116,7 @@ class GeneralElementDialog extends StatelessWidget {
               title: Text(AppLocalizations.of(context)!.duplicate),
               leading: const Icon(PhosphorIcons.copyLight),
               onTap: () {
-                editingCubit.move(element);
+                state.fetchHand()?.move(context, renderer, true);
                 close();
               },
             ),
@@ -143,48 +142,36 @@ class GeneralElementDialog extends StatelessWidget {
                                       Text(AppLocalizations.of(context)!.yes),
                                   onPressed: () {
                                     Navigator.pop(context);
-                                    bloc.add(ElementsRemoved([element]));
+                                    bloc.add(
+                                        ElementsRemoved([renderer.element]));
                                   },
                                 ),
                               ]));
                 },
                 title: Text(AppLocalizations.of(context)!.delete),
                 leading: const Icon(PhosphorIcons.trashLight)),
-            const Divider(),
-            ListTile(
-              title: Text(AppLocalizations.of(context)!.document),
-              leading: const Icon(PhosphorIcons.fileLight),
-              onTap: () async {
-                var bloc = context.read<DocumentBloc>();
-                var transformCubit = context.read<TransformCubit>();
-                var actor = context.findAncestorWidgetOfExactType<Actions>();
-                showContextMenu(
-                    context: context,
-                    position: position,
-                    builder: (ctx, close) => MultiBlocProvider(
-                          providers: [
-                            BlocProvider.value(value: bloc),
-                            BlocProvider.value(value: transformCubit),
-                          ],
-                          child: Actions(
-                              actions: actor?.actions ?? {},
-                              child: BackgroundContextMenu(
-                                close: close,
-                                position: position,
-                              )),
-                        ));
-                close();
-              },
-            ),
           ],
         ))
       ]);
     });
   }
+}
 
-  Widget generateHeader(PadElement element) {
+class _GeneralElementDialogHeader extends StatelessWidget {
+  final Renderer renderer;
+  final Offset position;
+  final VoidCallback close;
+  const _GeneralElementDialogHeader(
+      {Key? key,
+      required this.renderer,
+      this.position = Offset.zero,
+      required this.close})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
     IconData icon;
-    switch (element.toJson()['type']) {
+    switch (renderer.element.toJson()['type']) {
       case 'label':
         icon = PhosphorIcons.textTLight;
         break;
@@ -199,9 +186,70 @@ class GeneralElementDialog extends StatelessWidget {
         break;
     }
 
-    return SizedBox(
-      height: 50,
-      child: Center(child: Icon(icon, size: 36)),
+    var bloc = context.read<DocumentBloc>();
+    var transformCubit = context.read<TransformCubit>();
+    return Container(
+      height: 75,
+      padding: const EdgeInsets.symmetric(horizontal: 16),
+      child: Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
+        Padding(
+          padding: const EdgeInsets.all(4.0),
+          child: Icon(
+            icon,
+            size: 32,
+          ),
+        ),
+        Row(children: [
+          if (renderer.area != null)
+            IconButton(
+              tooltip: renderer.area!.name,
+              onPressed: () {
+                showContextMenu(
+                    context: context,
+                    position: position,
+                    builder: (context, close) => MultiBlocProvider(
+                            providers: [
+                              BlocProvider.value(value: bloc),
+                              BlocProvider.value(value: transformCubit),
+                            ],
+                            child: AreaContextMenu(
+                              area: renderer.area!,
+                              close: close,
+                              position: position,
+                            )));
+                close();
+              },
+              padding: const EdgeInsets.all(4.0),
+              icon: const Icon(PhosphorIcons.squareLight, size: 32),
+            ),
+          IconButton(
+            tooltip: AppLocalizations.of(context)!.document,
+            icon: const Icon(PhosphorIcons.fileLight, size: 32),
+            padding: const EdgeInsets.all(4.0),
+            onPressed: () async {
+              var bloc = context.read<DocumentBloc>();
+              var transformCubit = context.read<TransformCubit>();
+              var actor = context.findAncestorWidgetOfExactType<Actions>();
+              showContextMenu(
+                  context: context,
+                  position: position,
+                  builder: (ctx, close) => MultiBlocProvider(
+                        providers: [
+                          BlocProvider.value(value: bloc),
+                          BlocProvider.value(value: transformCubit),
+                        ],
+                        child: Actions(
+                            actions: actor?.actions ?? {},
+                            child: BackgroundContextMenu(
+                              close: close,
+                              position: position,
+                            )),
+                      ));
+              close();
+            },
+          ),
+        ]),
+      ]),
     );
   }
 }
