@@ -26,7 +26,7 @@ class _FileSystemAssetMoveDialogState extends State<FileSystemAssetMoveDialog> {
   void initState() {
     super.initState();
     _nameController.text = widget.asset.fileName;
-    selectedPath = widget.asset.directory;
+    selectedPath = widget.asset.parent;
   }
 
   @override
@@ -40,6 +40,7 @@ class _FileSystemAssetMoveDialogState extends State<FileSystemAssetMoveDialog> {
           TextButton(
             child: Text(AppLocalizations.of(context)!.ok),
             onPressed: () async {
+              final navigator = Navigator.of(context);
               var newPath = selectedPath;
               if (selectedPath != '/') {
                 newPath += '/';
@@ -50,7 +51,7 @@ class _FileSystemAssetMoveDialogState extends State<FileSystemAssetMoveDialog> {
               } else {
                 await _fileSystem.moveAsset(widget.asset.path, newPath);
               }
-              Navigator.of(context).pop(newPath);
+              navigator.pop(newPath);
             },
           ),
         ],
@@ -70,8 +71,8 @@ class _FileSystemAssetMoveDialogState extends State<FileSystemAssetMoveDialog> {
                         initialExpanded: true))),
             TextField(
               decoration: InputDecoration(
-                labelText: AppLocalizations.of(context)!.name,
-              ),
+                  filled: true, hintText: AppLocalizations.of(context)!.name),
+              autofocus: true,
               controller: _nameController,
             ),
           ],
