@@ -5,15 +5,13 @@ import 'dart:ui' as ui;
 import 'package:butterfly/bloc/document_bloc.dart';
 import 'package:butterfly/cubits/transform.dart';
 import 'package:butterfly/dialogs/camera.dart';
-import 'package:butterfly/models/elements/element.dart';
+import 'package:butterfly/models/element.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
-
-import '../models/elements/image.dart';
 
 class InsertDialog extends StatefulWidget {
   final Offset position;
@@ -29,12 +27,10 @@ class _InsertDialogState extends State<InsertDialog> {
     var state = bloc.state;
     var transform = context.read<TransformCubit>().state;
     if (state is! DocumentLoadSuccess) return;
-    var bakedViewport = state.bakedViewport;
+    var bakedViewport = state.cameraViewport;
     bloc.add(ElementsCreated([element]));
-    if (bakedViewport != null) {
-      bloc.add(ImageBaked(bakedViewport.toSize(),
-          MediaQuery.of(context).devicePixelRatio, transform));
-    }
+    bloc.add(ImageBaked(bakedViewport.toSize(),
+        MediaQuery.of(context).devicePixelRatio, transform));
     Navigator.of(context).pop();
   }
 
@@ -55,6 +51,7 @@ class _InsertDialogState extends State<InsertDialog> {
           title: Text(AppLocalizations.of(context)!.image),
           leading: const Icon(PhosphorIcons.imageLight),
           onTap: () async {
+            final bloc = context.read<DocumentBloc>();
             var files = await FilePicker.platform.pickFiles(
                 type: FileType.image, allowMultiple: false, withData: true);
             if (files?.files.isEmpty ?? true) return;
@@ -69,7 +66,6 @@ class _InsertDialogState extends State<InsertDialog> {
             var image = frame.image.clone();
 
             var bytes = await image.toByteData(format: ui.ImageByteFormat.png);
-            var bloc = context.read<DocumentBloc>();
             var state = bloc.state;
             if (state is! DocumentLoadSuccess) return;
             _submit(ImageElement(
@@ -88,6 +84,7 @@ class _InsertDialogState extends State<InsertDialog> {
               title: Text(AppLocalizations.of(context)!.camera),
               leading: const Icon(PhosphorIcons.cameraLight),
               onTap: () async {
+                final bloc = context.read<DocumentBloc>();
                 var content = await Navigator.of(context).push(
                   MaterialPageRoute(builder: (context) => const CameraDialog()),
                 ) as Uint8List?;
@@ -99,7 +96,6 @@ class _InsertDialogState extends State<InsertDialog> {
 
                 var bytes =
                     await image.toByteData(format: ui.ImageByteFormat.png);
-                var bloc = context.read<DocumentBloc>();
                 var state = bloc.state;
                 if (state is! DocumentLoadSuccess) return;
                 _submit(ImageElement(
