@@ -29,7 +29,7 @@ class LayersDialog extends StatelessWidget {
                     label:
                         Text(AppLocalizations.of(context)!.selectCustomLayer),
                     onPressed: () {
-                      var _nameController = TextEditingController(
+                      var nameController = TextEditingController(
                           text: (context.read<DocumentBloc>().state
                                   as DocumentLoadSuccess)
                               .currentLayer);
@@ -39,9 +39,11 @@ class LayersDialog extends StatelessWidget {
                           title:
                               Text(AppLocalizations.of(ctx)!.selectCustomLayer),
                           content: TextField(
-                            controller: _nameController,
+                            controller: nameController,
                             autofocus: true,
-                            decoration: const InputDecoration(filled: true),
+                            decoration: InputDecoration(
+                                filled: true,
+                                hintText: AppLocalizations.of(context)!.name),
                           ),
                           actions: [
                             TextButton(
@@ -53,7 +55,7 @@ class LayersDialog extends StatelessWidget {
                               onPressed: () {
                                 Navigator.of(ctx).pop();
                                 BlocProvider.of<DocumentBloc>(context).add(
-                                    CurrentLayerChanged(_nameController.text));
+                                    CurrentLayerChanged(nameController.text));
                               },
                             ),
                           ],
