@@ -1,13 +1,13 @@
 import 'package:butterfly/api/open_help.dart';
 import 'package:butterfly/bloc/document_bloc.dart';
 import 'package:butterfly/dialogs/color_pick.dart';
-import 'package:butterfly/models/painters/label.dart';
-import 'package:butterfly/models/properties/label.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 
+import '../../models/painter.dart';
+import '../../models/property.dart';
 import '../../widgets/exact_slider.dart';
 
 class LabelPainterDialog extends StatefulWidget {
@@ -146,9 +146,7 @@ class LabelPropertyView extends StatefulWidget {
   final LabelPropertyCallback onChanged;
 
   const LabelPropertyView(
-      {Key? key,
-      this.initialValue = const LabelProperty(),
-      required this.onChanged})
+      {Key? key, required this.initialValue, required this.onChanged})
       : super(key: key);
 
   @override
@@ -193,10 +191,52 @@ class _LabelPropertyViewState extends State<LabelPropertyView> {
           max: 20,
           onChanged: (value) => change(_value.copyWith(letterSpacing: value),
               value != _value.letterSpacing)),
+      ToggleButtons(
+          isSelected: [
+            _value.horizontalAlignment == HorizontalAlignment.left,
+            _value.horizontalAlignment == HorizontalAlignment.center,
+            _value.horizontalAlignment == HorizontalAlignment.right,
+            _value.horizontalAlignment == HorizontalAlignment.justify,
+          ],
+          onPressed: (index) {
+            var alignments = [
+              HorizontalAlignment.left,
+              HorizontalAlignment.center,
+              HorizontalAlignment.right,
+              HorizontalAlignment.justify,
+            ];
+            change(
+                _value.copyWith(horizontalAlignment: alignments[index]), true);
+          },
+          children: const [
+            Icon(PhosphorIcons.textAlignLeftLight),
+            Icon(PhosphorIcons.textAlignCenterLight),
+            Icon(PhosphorIcons.textAlignRightLight),
+            Icon(PhosphorIcons.textAlignJustifyLight),
+          ]),
+      ToggleButtons(
+          isSelected: [
+            _value.verticalAlignment == VerticalAlignment.top,
+            _value.verticalAlignment == VerticalAlignment.center,
+            _value.verticalAlignment == VerticalAlignment.bottom,
+          ],
+          onPressed: (index) {
+            var alignments = [
+              VerticalAlignment.top,
+              VerticalAlignment.center,
+              VerticalAlignment.bottom,
+            ];
+            change(_value.copyWith(verticalAlignment: alignments[index]), true);
+          },
+          children: const [
+            Icon(PhosphorIcons.alignTopLight),
+            Icon(PhosphorIcons.alignCenterVerticalLight),
+            Icon(PhosphorIcons.alignBottomLight),
+          ]),
       ListTile(
           title: Text(AppLocalizations.of(context)!.fontWeight),
           trailing: DropdownButton<FontWeight>(
-              value: _value.fontWeight,
+              value: FontWeight.values[_value.fontWeight],
               items: List.generate(FontWeight.values.length, (index) {
                 var text = ((index + 1) * 100).toString();
                 if (index == 3) {
@@ -205,14 +245,15 @@ class _LabelPropertyViewState extends State<LabelPropertyView> {
                   text = AppLocalizations.of(context)!.bold;
                 }
                 return DropdownMenuItem(
-                    child: Text(text), value: FontWeight.values[index]);
+                    value: FontWeight.values[index], child: Text(text));
               }),
-              onChanged: (value) =>
-                  change(_value.copyWith(fontWeight: value)))),
+              onChanged: (value) => change(_value.copyWith(
+                  fontWeight: value?.index ?? _value.fontWeight)))),
       CheckboxListTile(
           title: Text(AppLocalizations.of(context)!.italic),
           value: _value.italic,
-          onChanged: (value) => change(_value.copyWith(italic: value))),
+          onChanged: (value) =>
+              change(_value.copyWith(italic: value ?? _value.italic))),
       ExpansionPanelList(
         expansionCallback: (panelIndex, isExpanded) =>
             setState(() => _decorationExpanded = !_decorationExpanded),
@@ -232,18 +273,18 @@ class _LabelPropertyViewState extends State<LabelPropertyView> {
                 CheckboxListTile(
                     title: Text(AppLocalizations.of(context)!.lineThrough),
                     value: _value.lineThrough,
-                    onChanged: (value) =>
-                        change(_value.copyWith(lineThrough: value))),
+                    onChanged: (value) => change(_value.copyWith(
+                        lineThrough: value ?? _value.lineThrough))),
                 CheckboxListTile(
                     title: Text(AppLocalizations.of(context)!.underline),
                     value: _value.underline,
-                    onChanged: (value) =>
-                        change(_value.copyWith(underline: value))),
+                    onChanged: (value) => change(
+                        _value.copyWith(underline: value ?? _value.underline))),
                 CheckboxListTile(
                     title: Text(AppLocalizations.of(context)!.overline),
                     value: _value.overline,
-                    onChanged: (value) =>
-                        change(_value.copyWith(overline: value))),
+                    onChanged: (value) => change(
+                        _value.copyWith(overline: value ?? _value.overline))),
                 ListTile(
                     title: Text(AppLocalizations.of(context)!.style),
                     trailing: DropdownButton<TextDecorationStyle>(
@@ -269,11 +310,12 @@ class _LabelPropertyViewState extends State<LabelPropertyView> {
                               text = AppLocalizations.of(context)!.wavy;
                           }
                           return DropdownMenuItem(
-                              child: Text(text),
-                              value: TextDecorationStyle.values[index]);
+                              value: TextDecorationStyle.values[index],
+                              child: Text(text));
                         }),
-                        onChanged: (value) =>
-                            change(_value.copyWith(decorationStyle: value)))),
+                        onChanged: (value) => change(_value.copyWith(
+                            decorationStyle:
+                                value ?? _value.decorationStyle)))),
                 ListTile(
                     onTap: () async {
                       var value = await showDialog(
@@ -281,18 +323,18 @@ class _LabelPropertyViewState extends State<LabelPropertyView> {
                           builder: (ctx) => BlocProvider.value(
                                 value: context.read<DocumentBloc>(),
                                 child: ColorPickerDialog(
-                                    defaultColor: _value.decorationColor),
-                              ));
+                                    defaultColor:
+                                        Color(_value.decorationColor)),
+                              )) as Color?;
                       if (value != null) {
-                        change(
-                            _value.copyWith(decorationColor: value as Color));
+                        change(_value.copyWith(decorationColor: value.value));
                       }
                     },
                     leading: Container(
                         width: 30,
                         height: 30,
                         decoration: BoxDecoration(
-                            color: _value.decorationColor,
+                            color: Color(_value.decorationColor),
                             borderRadius:
                                 const BorderRadius.all(Radius.circular(32)))),
                     title: Text(AppLocalizations.of(context)!.color)),
@@ -321,7 +363,8 @@ class _LabelPropertyViewState extends State<LabelPropertyView> {
                     context: context,
                     builder: (ctx) => BlocProvider.value(
                           value: context.read<DocumentBloc>(),
-                          child: ColorPickerDialog(defaultColor: _value.color),
+                          child: ColorPickerDialog(
+                              defaultColor: Color(_value.color)),
                         ));
                 if (color != null) {
                   change(_value.copyWith(color: color));
@@ -329,7 +372,7 @@ class _LabelPropertyViewState extends State<LabelPropertyView> {
               },
               child: Container(
                   decoration: BoxDecoration(
-                      color: _value.color,
+                      color: Color(_value.color),
                       borderRadius:
                           const BorderRadius.all(Radius.circular(32))),
                   constraints:
