@@ -1,11 +1,11 @@
 import 'package:butterfly/api/open_help.dart';
 import 'package:butterfly/bloc/document_bloc.dart';
-import 'package:butterfly/models/painters/path_eraser.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 
+import '../../models/painter.dart';
 import '../../widgets/exact_slider.dart';
 
 class PathEraserPainterDialog extends StatefulWidget {
@@ -80,7 +80,9 @@ class _PathEraserPainterDialogState extends State<PathEraserPainterDialog> {
                                 title: Text(AppLocalizations.of(context)!
                                     .includeEraser),
                                 onChanged: (value) => setState(() => painter =
-                                    painter.copyWith(includeEraser: value))),
+                                    painter.copyWith(
+                                        includeEraser:
+                                            value ?? painter.includeEraser))),
                             /*CheckboxListTile(
                                 value: painter.deleteWholeStroke,
                                 title: Text(AppLocalizations.of(context)!
