@@ -1,78 +1,77 @@
-import 'package:butterfly/cubits/selection.dart';
-import 'package:butterfly/models/elements/element.dart';
+import 'package:butterfly/renderers/renderer.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 
-class SelectElementDialog extends StatelessWidget {
-  final List<PadElement> elements;
-  final SelectionCubit cubit;
+class SelectElementDialog extends StatefulWidget {
+  final List<Renderer> renderers;
 
-  const SelectElementDialog(
-      {Key? key, this.elements = const [], required this.cubit})
+  const SelectElementDialog({Key? key, this.renderers = const []})
       : super(key: key);
 
   @override
+  State<SelectElementDialog> createState() => _SelectElementDialogState();
+}
+
+class _SelectElementDialogState extends State<SelectElementDialog> {
+  Renderer? current;
+
+  @override
+  void initState() {
+    current = widget.renderers.isEmpty ? null : widget.renderers.first;
+    super.initState();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    return BlocProvider.value(
-      value: cubit,
-      child: AlertDialog(
+    return AlertDialog(
         title: Text(AppLocalizations.of(context)!.selectElement),
         actions: [
           TextButton(
             child: Text(AppLocalizations.of(context)!.cancel),
             onPressed: () {
-              cubit.reset();
               Navigator.of(context).pop();
             },
           ),
           TextButton(
             child: Text(AppLocalizations.of(context)!.ok),
-            onPressed: () => Navigator.of(context).pop(true),
+            onPressed: () => Navigator.of(context).pop(current),
           ),
         ],
         content: ConstrainedBox(
           constraints: const BoxConstraints(maxWidth: 100),
-          child:
-              BlocBuilder<SelectionCubit, dynamic>(builder: (context, state) {
-            return SizedBox(
-                width: double.maxFinite,
-                child: ListView.builder(
-                  shrinkWrap: true,
-                  itemCount: elements.length,
-                  itemBuilder: (context, index) {
-                    final element = elements[index];
-                    final elementType = element.toJson()['type'];
-                    IconData icon;
-                    switch (elementType) {
-                      case 'image':
-                        icon = PhosphorIcons.imageLight;
-                        break;
-                      case 'label':
-                        icon = PhosphorIcons.textTLight;
-                        break;
-                      case 'eraser':
-                        icon = PhosphorIcons.eraserLight;
-                        break;
-                      default:
-                        icon = PhosphorIcons.penLight;
-                        break;
-                    }
-                    return IconButton(
-                      icon: Icon(icon),
-                      color: state == element
-                          ? Theme.of(context).colorScheme.primary
-                          : null,
-                      onPressed: () {
-                        cubit.change(element);
-                      },
-                    );
-                  },
-                ));
-          }),
-        ),
-      ),
-    );
+          child: SizedBox(
+            width: double.maxFinite,
+            child: ListView.builder(
+                shrinkWrap: true,
+                itemCount: widget.renderers.length,
+                itemBuilder: (context, index) {
+                  final element = widget.renderers[index];
+                  final elementType = element.element.toJson()['type'];
+                  IconData icon;
+                  switch (elementType) {
+                    case 'image':
+                      icon = PhosphorIcons.imageLight;
+                      break;
+                    case 'label':
+                      icon = PhosphorIcons.textTLight;
+                      break;
+                    case 'eraser':
+                      icon = PhosphorIcons.eraserLight;
+                      break;
+                    default:
+                      icon = PhosphorIcons.penLight;
+                      break;
+                  }
+                  return IconButton(
+                    icon: Icon(icon),
+                    color: current == element
+                        ? Theme.of(context).colorScheme.primary
+                        : null,
+                    onPressed: () => setState(() => current = element),
+                  );
+                }),
+          ),
+        ));
   }
 }
