@@ -7,7 +7,7 @@ import 'package:butterfly/api/full_screen_stub.dart'
     if (dart.library.js) 'package:butterfly/api/full_screen_html.dart'
     as full_screen;
 
-import 'models/document.dart';
+import 'models/converter.dart';
 
 Future<void> setup() async {
   // Convert old file system to new file system
@@ -19,7 +19,8 @@ Future<void> setup() async {
     var documents = prefs.getStringList('documents')!;
     prefs.remove('documents');
     await Future.wait(documents
-        .map((e) => AppDocument.fromJson(jsonDecode(e)))
+        .map((e) => const DocumentJsonConverter()
+            .fromJson(Map<String, dynamic>.from(jsonDecode(e))))
         .map((element) => DocumentFileSystem.fromPlatform()
             .updateDocument(element.name, element)));
   }
